Update github action version
Some actions can be potentially problematic to use in the future because of
deprecation. For example, The actions/setup-go@v2 uses node12 which is deprecated and will be forced to run on node16.
From github.blog:
```
To opt out of this and continue using Node12 while it is still available in the runner, you can choose to set ACTIONS_ALLOW_USE_UNSECURE_NODE_VERSION=true
as an 'env' in their workflow or as an environment variable on your runner machine. This will only work until we upgrade the runner removing Node12 later in the summer.
```